spec-364: split ui/api/client.ts into per-domain modules behind a barrel [P3]
Behaviour-preserving carve of the ~2,866-line all-domains api/client.ts into
per-domain modules (docs, search, comments, decisions, tasks, issues, auth, org,
memex, mcp, drift, integrations, oauth, acs, insights, billing) plus internal.ts
(tBase, orgBillingBase, OrgTenant). client.ts is now a thin barrel that
`export *`s every domain module and keeps the original errors/fetchJson/http +
Tag/MemexIssue/NamespaceHomeResponse/MemexDto/OrgTenant re-exports, so all ~185
`from "../api/client"` imports resolve unchanged — same names, signatures,
return types. No contract change; only packages/ui/src/api/* touched.

Finding sol-2 of audit spec-345, under decomposition umbrella spec-354.

Verified green: pnpm --filter @memex/ui exec tsc -b; make test-ui
(213 files, 1508 pass / 1 skip); pnpm --filter @memex/ui build.
